requirements.txt
pytest
allure-pytest
appium-python-client
aiogram
aiohttp
